Foi implementado a interface de recarga do cartão no fornecedor agilitas, bem como disponibiização na interface de api da aplicação

+++ server/allowpay-soluction/allowpay-api/src/main/java/br/com/allowpay/integrator/converters/ExtractDtoToFullExtractConverter.java
@@ -10,7 +10,7 @@ import br.com.allowpay.integrator.dtos.ExtractValuesDto;
 public class ExtractDtoToFullExtractConverter {
 	
 	public Extract convert(final ExtractValuesDto extractDto){
-		final Integer value = extractDto.getValor();
+		final Double value = extractDto.getValor();
 		final String date = extractDto.getDataHora();
 		final String merchant = extractDto.getEstabelecimento();
 		final String type = extractDto.getTipo();
